requirements.txt
requests
pytest
pyyaml
pytest-assume